@@ -37,13 +37,14 @@
     "lint:fix": "eslint --ext=.ts,tsx src --fix",
     "stylelint": "stylelint \"src/**/*.styled.{tsx,ts}\"",
     "prepare": "husky install .husky",
-    "storybook": "start-storybook -p 6007 -s public",
-    "build-storybook": "build-storybook --quiet",
+    "start-torybook-server": "http-server ./storybook-static -p 6006 -s",
+    "storybook": "storybook dev -p 6007",
+    "build-storybook": "storybook build --quiet",
     "test-storybook": "test-storybook",
     "test": "jest test",
     "test:ci": "cross-env CI=true jest test --passWithNoTests --testResultsProcessor=jest-teamcity-reporter",
     "creevey:ui": "creevey --ui --port=3003",
-    "screenshots:ci": "concurrently -k -s first \"http-server ./storybook-static -p 6006 -s\" \"yarn creevey\" \"yarn test-storybook\""
+    "screenshots:ci": "concurrently -k -s first \"yarn start-storybook-server\" \"yarn creevey\" \"yarn test-storybook\""
   },
   "peerDependencies": {
     "@skbkontur/react-ui": ">=4.5.0 <5",
@@ -52,34 +53,30 @@
     "react-dom": ">=16.9"
   },
   "devDependencies": {
-    "@babel/core": "7.18.0",
-    "@babel/plugin-syntax-jsx": "7.21.4",
-    "@babel/preset-env": "7.21.4",
-    "@babel/preset-typescript": "7.21.4",
     "@skbkontur/react-ui": "^4.14.0",
     "@skbkontur/react-ui-validations": "^1.12.0",
-    "@storybook/addon-a11y": "6.5.16",
-    "@storybook/addon-actions": "6.5.16",
-    "@storybook/addon-essentials": "6.5.16",
-    "@storybook/addon-interactions": "6.5.3",
-    "@storybook/addon-links": "6.5.16",
-    "@storybook/builder-webpack5": "6.5.16",
-    "@storybook/manager-webpack5": "6.5.16",
-    "@storybook/node-logger": "6.5.16",
-    "@storybook/preset-create-react-app": "4.1.2",
-    "@storybook/react": "6.5.16",
-    "@storybook/test-runner": "0.9.4",
+    "@storybook/addon-a11y": "8.2.8",
+    "@storybook/addon-actions": "8.2.8",
+    "@storybook/addon-essentials": "8.2.8",
+    "@storybook/addon-interactions": "8.2.8",
+    "@storybook/addon-links": "8.2.8",
+    "@storybook/node-logger": "8.2.8",
+    "@storybook/react": "8.2.8",
+    "@storybook/react-vite": "8.2.8",
+    "@storybook/test-runner": "0.19.1",
     "@swc/core": "1.3.50",
     "@swc/jest": "0.2.26",
+    "@swc/plugin-styled-components": "5.0.0",
     "@types/jest": "29.5.0",
     "@types/node": "12.0.0",
-    "@types/react": "16.9.0",
-    "@types/react-dom": "16.9.0",
+    "@types/react": "17.0.39",
+    "@types/react-dom": "17.0.25",
     "@typescript-eslint/eslint-plugin": "5.57.0",
     "@typescript-eslint/parser": "5.57.0",
+    "@vitejs/plugin-react-swc": "3.7.2",
     "axe-playwright": "1.2.3",
     "concurrently": "^7.0.0",
-    "creevey": "0.9.0-beta.5",
+    "creevey": "0.9.2",
     "cross-env": "6.0.3",
     "eslint": "8.32.0",
     "eslint-config-prettier": "8.8.0",
@@ -98,13 +95,15 @@
     "lerna": "5.0.0",
     "lint-staged": "12.3.4",
     "prettier": "2.2.1",
-    "react": "16.9.0",
-    "react-dom": "16.9.0",
+    "react": "17",
+    "react-dom": "17",
+    "storybook": "8.2.8",
     "stylelint": "13.13.1",
     "stylelint-config-standard": "22.0.0",
     "stylelint-config-styled-components": "^0.1.1",
     "stylelint-processor-styled-components": "^1.10.0",
-    "typescript": "4.9.4"
+    "typescript": "4.9.4",
+    "vite": "5.4.11"
   },
   "files": [
     "src/**/*.js",
