@@ -1,27 +1,26 @@
 import { injectAxe, checkA11y, configureAxe } from 'axe-playwright';
 
-import { getStoryContext } from '@storybook/test-runner';
+import { getStoryContext, TestRunnerConfig } from '@storybook/test-runner';
+import { ConfigOptions } from 'axe-playwright/dist/types';
 
-/*
- * See https://storybook.js.org/docs/7.0/react/writing-tests/test-runner#test-hook-api-experimental
- * to learn more about the test-runner hooks API.
- */
-const a11yConfig = {
-  async preRender(page) {
+const a11yConfig: TestRunnerConfig = {
+  async preVisit(page) {
     await injectAxe(page);
   },
-  async postRender(page, context) {
+  async postVisit(page, context) {
     // Get the entire context of a story, including parameters, args, argTypes, etc.
     const storyContext = await getStoryContext(page, context);
 
     // Do not test a11y for stories that disable a11y
-    if (storyContext.parameters?.a11y?.disable) return;
+    if (storyContext.parameters?.a11y?.disable) {
+      return;
+    }
 
     // Apply story-level a11y rules
     await configureAxe(page, {
       rules: storyContext.parameters?.a11y?.config?.rules ?? [],
       disableOtherRules: true,
-    });
+    } as ConfigOptions);
 
     await checkA11y(page, '#test-element', {
       detailedReport: true,
@@ -29,9 +28,8 @@ const a11yConfig = {
         html: true,
       },
       // pass axe options defined in @storybook/addon-a11y
-      axeOptions: storyContext.parameters?.a11y?.options,
     });
   },
 };
 
-module.exports = a11yConfig;
+export default a11yConfig;
