@@ -1,12 +1,14 @@
 import { Modal, SizeProp, THEME_2022 } from '@skbkontur/react-ui';
 import { text, ValidationContainer } from '@skbkontur/react-ui-validations';
-import { Meta, Story } from '@storybook/react';
+import { Meta, StoryFn } from '@storybook/react';
 import React, { useState } from 'react';
 
 import { a11yRules } from '../../../a11y/rules';
-import { MarkdownApi, MarkdownThemeProvider, MarkdownViewer, RefItem, User } from '../../../index';
+import { MarkdownViewer } from '../../MarkdownViewer';
+import { MarkdownThemeProvider } from '../../styles/theme';
 import { allVariantsMarkdownMock } from '../__mocks__/markdown.mock';
 import { Markdown } from '../Markdown';
+import { MarkdownApi, RefItem, User } from '../types';
 
 export default {
   title: 'Markdown',
@@ -29,7 +31,7 @@ const apiMock: MarkdownApi = {
 
 export const WithoutActions = () => <Markdown hideMarkdownActions value={allVariantsMarkdownMock} />;
 
-export const WithSizeControl: Story<{ size: SizeProp }> = args => (
+export const WithSizeControl: StoryFn<{ size: SizeProp }> = args => (
   <Markdown size={args.size} value={allVariantsMarkdownMock} />
 );
 
@@ -75,7 +77,7 @@ export const WithValidation = () => {
   );
 };
 
-export const InModal: Story = () => {
+export const InModal: StoryFn = () => {
   const [value, setValue] = useState<string>('');
 
   return (
@@ -107,7 +109,7 @@ export const InModal: Story = () => {
 
 InModal.parameters = { creevey: { captureElement: 'body > div.react-ui > div' } };
 
-export const CustomWidth: Story = () => {
+export const CustomWidth: StoryFn = () => {
   const [value, setValue] = useState<string>(allVariantsMarkdownMock);
 
   return <Markdown width="550px" fileApiUrl="/api/file" value={value} onValueChange={setValue} />;
@@ -115,7 +117,7 @@ export const CustomWidth: Story = () => {
 
 CustomWidth.decorators = [];
 
-export const CustomValidation: Story = () => {
+export const CustomValidation: StoryFn = () => {
   const [value, setValue] = useState<string>(allVariantsMarkdownMock);
 
   return (
@@ -132,7 +134,7 @@ export const CustomValidation: Story = () => {
   );
 };
 
-export const WithoutHints: Story = () => {
+export const WithoutHints: StoryFn = () => {
   return (
     <ValidationContainer>
       <Markdown
@@ -146,8 +148,12 @@ export const WithoutHints: Story = () => {
   );
 };
 
-export const Viewer = () => (
+export const Viewer: StoryFn = () => (
   <div style={{ width: 320 }}>
     <MarkdownViewer fileApiUrl="/api/file/download" source={allVariantsMarkdownMock} />
   </div>
 );
+
+Viewer.parameters = {
+  creevey: { delay: 5000 },
+};
